+++ src/main/java/com/zylitics/zwl/webdriver/functions/action/DragAndDrop.java
@@ -1,5 +1,6 @@
 package com.zylitics.zwl.webdriver.functions.action;
 
+import com.google.common.io.Resources;
 import com.zylitics.zwl.datatype.Types;
 import com.zylitics.zwl.webdriver.TimeoutType;
 import com.zylitics.zwl.webdriver.functions.AbstractWebdriverFunction;
@@ -7,15 +8,22 @@ import com.zylitics.zwl.datatype.ZwlValue;
 import com.zylitics.zwl.webdriver.APICoreProperties;
 import com.zylitics.zwl.webdriver.BuildCapability;
 import org.openqa.selenium.InvalidElementStateException;
+import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.remote.RemoteWebElement;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
+import java.io.IOException;
 import java.io.PrintStream;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
+import java.util.Objects;
 import java.util.function.Supplier;
 
+// TODO: Currently only the 2 args variant works reliably for HTML5 dnd. Either remove the other variants
+//  or see if they can be done using the js.
 public class DragAndDrop extends AbstractWebdriverFunction {
   
   public DragAndDrop(APICoreProperties.Webdriver wdProps,
@@ -63,7 +71,10 @@ public class DragAndDrop extends AbstractWebdriverFunction {
             "waiting for element to become intractable");
         wait.until(d -> {
           try {
-            actions.dragAndDrop(el1, el2);
+            driver.executeScript(
+                String.format("%s\ndragAndDrop(arguments[0], arguments[1]);",
+                    getAtom()),
+                el1, el2);
             return true;
           } catch (InvalidElementStateException ie) {
             return false;
@@ -90,5 +101,17 @@ public class DragAndDrop extends AbstractWebdriverFunction {
   @Override
   protected String getFuncReturnType() {
     return Types.VOID;
+  }
+  
+  private String getAtom() {
+    try {
+      String scriptName = "/dragAndDrop.min.js";
+      URL url = getClass().getResource(scriptName);
+      Objects.requireNonNull(url);
+      //noinspection UnstableApiUsage
+      return Resources.toString(url, StandardCharsets.UTF_8);
+    } catch (IOException | NullPointerException e) {
+      throw new WebDriverException(e);
+    }
   }
 }